backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
pymongo==4.6.2
python-crontab==3.0.0
motor>=3.0.0
yfinance==0.2.36
pandas
apscheduler
ollama
sentence-transformers
langchain-core
langchain-text-splitters
faiss-cpu
numpy
jinja2